chore(types): remove redundant transform.ts
